Deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
numpy
plotly
pillow
scikit-learn==1.3.2